Rename package to com/contrastsecurity/maven/plugin
